fix(storage): revert moving location credentials provider interface
